[UPDATE] 수정사항 8,9번 코스 visibility, editable 유무 반영

+++ src/main/java/com/dalc/one/controller/CourseController.java
@@ -42,7 +42,25 @@ public class CourseController{
 	@ResponseBody
 	@GetMapping
 	public ResponseEntity<Course> getCourse(HttpServletRequest request, @RequestParam("cid") int id) throws Exception {
-		return ResponseEntity.ok(lehgo.getCourse(id));
+		Course course = lehgo.getCourse(id);
+		String authorizationHeader = request.getHeader("authorization");
+
+		if(course.getVisibility() == 1) { // 작성자만 view 가능할 때
+			if (!JwtTokenProvider.getUserOf(authorizationHeader).getUsername().equals(course.getUserId())) { // 코스의 ID와 로그인 ID가 일치하지 않으면 오류
+				throw new ResponseStatusException
+					(ExceptionEnum.PERMISSION_FAIL.getStatus(), ExceptionEnum.PERMISSION_FAIL.getMessage());
+			}
+			else {
+				return ResponseEntity.ok(lehgo.getCourse(id));
+			}
+		}
+			return ResponseEntity.ok(lehgo.getCourse(id));
+	}
+	
+	@ResponseBody
+	@GetMapping("all")
+	public ResponseEntity<List<Course>> getVisibleCourseList(HttpServletRequest request) throws Exception {
+		return ResponseEntity.ok(lehgo.getVisibleCourse());
 	}
 	
 	@ResponseBody
@@ -79,6 +97,18 @@ public class CourseController{
 	@ResponseBody
 	@GetMapping("detail")
 	public ResponseEntity<List<CoursePlace>> getCourseDetail(HttpServletRequest request, @RequestParam("cid") int courseId) throws Exception {	
+		Course course = lehgo.getCourse(courseId);
+		String authorizationHeader = request.getHeader("authorization");
+
+		if(course.getVisibility() == 1) { // 작성자만 view 가능할 때
+			if (!JwtTokenProvider.getUserOf(authorizationHeader).getUsername().equals(course.getUserId())) { // 코스의 ID와 로그인 ID가 일치하지 않으면 오류
+				throw new ResponseStatusException
+					(ExceptionEnum.PERMISSION_FAIL.getStatus(), ExceptionEnum.PERMISSION_FAIL.getMessage());
+			}
+			else // 코스의 작성자면 view
+				return ResponseEntity.ok(lehgo.getCourseDetail(courseId));
+		} 
+		// 접속한 모든 사람들 view 가능
 		return ResponseEntity.ok(lehgo.getCourseDetail(courseId));
 	}
 	
@@ -132,19 +162,34 @@ public class CourseController{
 	@ResponseBody
 	@PutMapping("detail/update")
 	public ResponseEntity<CoursePlace> updateCourseDetail(HttpServletRequest request, @RequestBody CoursePlace coursePlace) throws Exception {
-		CoursePlace result = lehgo.updateCoursePlace(coursePlace);
-		
-		if (result == null) {
-			throw new ResponseStatusException
-				(ExceptionEnum.INPUT_FAIL.getStatus(), ExceptionEnum.INPUT_FAIL.getMessage());
+		Course course = lehgo.getCourse(coursePlace.getCourseId());
+		CoursePlace result = new CoursePlace();
+		String authorizationHeader = request.getHeader("authorization");
+
+		if(course.getEditable() == 1) { // 작성자만 edit 가능할 때
+			if (!JwtTokenProvider.getUserOf(authorizationHeader).getUsername().equals(course.getUserId())) { // 코스의 ID와 로그인 ID가 일치하지 않으면 오류
+				throw new ResponseStatusException
+					(ExceptionEnum.NOT_MATCH.getStatus(), ExceptionEnum.NOT_MATCH.getMessage());
+			}
+			else {
+				result = lehgo.updateCoursePlace(coursePlace);
+			}
 		}
+		else { // 로그인한 모두가 edit 가능
+			if (authorizationHeader == null) {
+				throw new ResponseStatusException
+					(ExceptionEnum.NOT_LOGIN.getStatus(), ExceptionEnum.NOT_LOGIN.getMessage());
+			}
+			else
+				result = lehgo.updateCoursePlace(coursePlace);
+		}
+		
 		return ResponseEntity.ok(result);
 	}
 	
 	@ResponseBody
 	@DeleteMapping("delete")
 	public ResponseEntity<HttpStatus> deleteCourse(HttpServletRequest request, @RequestParam("cid") int cid) throws Exception {
-		
 		int result = lehgo.deleteCourse(cid);
 
 		if(result > 0) return ResponseEntity.ok(HttpStatus.OK);
@@ -154,9 +199,28 @@ public class CourseController{
 	@ResponseBody
 	@DeleteMapping("detail/delete")
 	public ResponseEntity<HttpStatus> deleteCourseDetail(HttpServletRequest request, @RequestParam("cid") int cid, @RequestParam("pid") int pid) throws Exception {
-		
-		int result = lehgo.deleteCourseDetail(cid, pid);
+		Course course = lehgo.getCourse(cid);
+		String authorizationHeader = request.getHeader("authorization");
 
+		int result = 0;
+		if(course.getEditable() == 1) { // 작성자만 edit 가능할 때
+			if (!JwtTokenProvider.getUserOf(authorizationHeader).getUsername().equals(course.getUserId())) { // 코스의 ID와 로그인 ID가 일치하지 않으면 오류
+				throw new ResponseStatusException
+					(ExceptionEnum.NOT_MATCH.getStatus(), ExceptionEnum.NOT_MATCH.getMessage());
+			}
+			else {
+				result = lehgo.deleteCourseDetail(cid, pid);
+			}
+		}
+		else { // 로그인한 모두가 edit 가능
+			if (authorizationHeader == null) {
+				throw new ResponseStatusException
+					(ExceptionEnum.NOT_LOGIN.getStatus(), ExceptionEnum.NOT_LOGIN.getMessage());
+			}
+			else
+				result = lehgo.deleteCourseDetail(cid, pid);
+		}
+		
 		if(result > 0) return ResponseEntity.ok(HttpStatus.OK);
 		else return ResponseEntity.ok(HttpStatus.CONFLICT);
 	}
